Refactor: remove refactor folder
