+++ scanapi/evaluators/__init__.py
@@ -1,0 +1,5 @@
+from scanapi.evaluators.code_evaluator import CodeEvaluator
+from scanapi.evaluators.spec_evaluator import SpecEvaluator
+from scanapi.evaluators.string_evaluator import StringEvaluator
+
+__all__ = ["CodeEvaluator", "SpecEvaluator", "StringEvaluator"]
