--- scanapi/refactor/evaluators/__init__.py
@@ -1,5 +1,0 @@
-from scanapi.refactor.evaluators.code_evaluator import CodeEvaluator
-from scanapi.refactor.evaluators.spec_evaluator import SpecEvaluator
-from scanapi.refactor.evaluators.string_evaluator import StringEvaluator
-
-__all__ = ["CodeEvaluator", "SpecEvaluator", "StringEvaluator"]
