+++ scanapi/utils.py
